+++ lucene/queries/src/java/org/apache/lucene/queries/function/valuesource/FloatVectorValueSource.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.lucene.queries.function.valuesource.densevectors;
+package org.apache.lucene.queries.function.valuesource;
 
 import java.io.IOException;
 import java.util.Arrays;
