+++ lucene/queries/src/test/org/apache/lucene/queries/function/TestKnnVectorSimilarityFunctions.java
@@ -27,12 +27,12 @@ import org.apache.lucene.document.StringField;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.index.IndexWriterConfig;
 import org.apache.lucene.index.VectorSimilarityFunction;
-import org.apache.lucene.queries.function.valuesource.densevectors.ByteVectorSimilarityFunction;
-import org.apache.lucene.queries.function.valuesource.densevectors.ByteVectorValueSource;
-import org.apache.lucene.queries.function.valuesource.densevectors.ByteVectorFieldSource;
-import org.apache.lucene.queries.function.valuesource.densevectors.FloatVectorValueSource;
-import org.apache.lucene.queries.function.valuesource.densevectors.FloatVectorFieldSource;
-import org.apache.lucene.queries.function.valuesource.densevectors.FloatVectorSimilarityFunction;
+import org.apache.lucene.queries.function.valuesource.ByteVectorSimilarityFunction;
+import org.apache.lucene.queries.function.valuesource.ByteVectorValueSource;
+import org.apache.lucene.queries.function.valuesource.ByteVectorFieldSource;
+import org.apache.lucene.queries.function.valuesource.FloatVectorValueSource;
+import org.apache.lucene.queries.function.valuesource.FloatVectorFieldSource;
+import org.apache.lucene.queries.function.valuesource.FloatVectorSimilarityFunction;
 import org.apache.lucene.search.IndexSearcher;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.search.ScoreDoc;
